refactor: 💡 dashboard builder
